backend/requirements.txt
langchain
langchain-groq
langgraph
tavily-python
yfinance
fastapi
uvicorn
python-dotenv
langchain-community
streamlit
reportlab